requirements.txt
alembic==1.0.11
amqp==2.5.2
audioread==2.1.8
autopep8==1.4.4
Babel==2.7.0
bcrypt==3.1.7
billiard==3.6.2.0
blinker==1.4
cffi==1.12.3
Click==7.0
colour==0.1.5
cycler==0.10.0
decorator==4.4.1
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Executor==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
flask-reverse-proxy-fix==0.2.1
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
importlib-metadata==1.5.0
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==0.14.1
kiwisolver==1.1.0
kombu==4.6.7
librosa==0.7.2
llvmlite==0.31.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.3.4
numba==0.47.0
numpy==1.17.4
passlib==1.7.1
psycopg2==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pydub==0.23.1
pyparsing==2.4.5
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
resampy==0.2.2
scikit-learn==0.22.1
scipy==1.5.4
six==1.12.0
SoundFile==0.10.3.post1
speaklater==1.3
SQLAlchemy==1.3.6
SQLAlchemy-Utils==0.35.0
termcolor==1.1.0
tqdm==4.40.1
validators==0.14.1
vine==1.3.0
Werkzeug==0.15.5
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
zipp==3.0.0
